// AngularMain/src/app/modules/models/UserFunctions/UserFunctionDto.ts
    export interface UserFunctionDto {
        tenentID: number;
        locatioN_ID: number;
        useR_ID: number;
        rolE_ID: number;
        menU_ID: number;
        masteR_ID: number;
        modulE_ID: number;
        menU_TYPE: string;
        menU_NAMEEnglish: string;
        menU_NAMEArabic: string;
        smalltext: string;
        fulL_NAME: string;
        logiN_ID: string;
        password: string;
        useR_TYPE: number;
        activE_FLAG: number;
        activetilldate: Date;
        link: string;
        urloption: string;
        urlrewrite: string;
        menU_LOCATION: string;
        menU_ORDER: number;
        doC_PARENT: string;
        addflage: number;
        editflage: number;
        delflage: number;
        printflage: number;
        amiglobale: number;
        mypersonal: number;
        iconpath: string;
        sP1: number;
        sP2: number;
        sP3: number;
        sP4: number;
        sP5: number;
        sP1Name: string;
        sP2Name: string;
        sP3Name: string;
        sP4Name: string;
        sP5Name: string;
        uploadDate: Date;
        uploadby: string;
        syncDate: Date;
        syncby: string;
        synID: number;
        cruP_ID: number;
        checked?: boolean;
        menu_Level: number;
    }